refactor ngen request response class into ngen request module
