+++ python/lib/communication/dmod/communication/maas_request/ngen/ngen_request.py
@@ -528,3 +528,41 @@ class NGENRequest(ModelExecRequest):
             model["partition_config_data_id"] = self.partition_cfg_data_id
 
         return {"model": model, "session-secret": self.session_secret}
+
+
+class NGENRequestResponse(ModelExecRequestResponse):
+    """
+    A response to a :class:`NGENRequest`.
+
+    Note that, when not ``None``, the :attr:`data` value will be a dictionary with the following format:
+        - key 'job_id' : the appropriate job id value in response to the request
+        - key 'scheduler_response' : the related :class:`SchedulerRequestResponse`, in serialized dictionary form
+
+    For example:
+    {
+        'job_id': 1,
+        'output_data_id': '00000000-0000-0000-0000-000000000000',
+        'scheduler_response': {
+            'success': True,
+            'reason': 'Testing Stub',
+            'message': 'Testing stub',
+            'data': {
+                'job_id': 1
+            }
+        }
+    }
+
+    Or:
+    {
+        'job_id': 0,
+        'output_data_id': '00000000-0000-0000-0000-000000000000',
+        'scheduler_response': {
+            'success': False,
+            'reason': 'Testing Stub',
+            'message': 'Testing stub',
+            'data': {}
+        }
+    }
+    """
+
+    response_to_type = NGENRequest
